pyproject.toml
[project]
name = "fortiface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "mediapipe>=0.10.20",
    "pydantic-settings>=2.7.1",
    "pymilvus>=2.5.4",
]
